chore: move `ServerItem` to `containers` folder
